refactor: move Color enum to utility package
