+++ src/main/java/com/codenames/codenames_backend/playingfield/Board.java
@@ -1,5 +1,6 @@
 package com.codenames.codenames_backend.playingfield;
 
+import com.codenames.codenames_backend.utility.Color;
 import java.util.List;
 import lombok.Getter;
