+++ src/main/java/com/codenames/codenames_backend/utility/Color.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames_backend.playingfield;
+package com.codenames.codenames_backend.utility;
 
 /**
  * Enum representing the color of a card on the board.
