+++ src/test/java/com/codenames/codenames_backend/playingfield/CardTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.codenames.codenames_backend.utility.Color;
 import org.junit.jupiter.api.Test;
 
 /** Unit tests for Card. */
