+++ src/test/java/com/codenames/codenames_backend/serialization/DataTransferObjectServiceTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.codenames.codenames_backend.playingfield.Card;
-import com.codenames.codenames_backend.playingfield.Color;
+import com.codenames.codenames_backend.utility.Color;
 import com.codenames.codenames_backend.playingfield.GameManager;
 import com.codenames.codenames_backend.serialization.DataTransferObjectService.Role;
 import java.util.List;
